+++ src/wrench/exceptions.py
@@ -16,3 +16,8 @@ class HttpRequestError(WrenchError):
 
 class FingerprintMismatchError(WrenchError):
     ...
+
+
+class ImportParseError(WrenchError):
+    def __init__(self, lineno):
+        self.lineno = lineno